Move TodoList into it's own component
