+++ src/todos/components/MainSection.js
@@ -1,12 +1,16 @@
 import React, { Component, PropTypes } from 'react';
+import { bindActionCreators } from 'redux';
+import { connect } from 'react-redux';
 
 import { Checkbox, List } from 'material-ui';
 import { getMuiTheme } from 'material-ui/styles';
-import MyRawTheme from '../material_ui_raw_theme_file';
+import MyRawTheme from '../../material_ui_raw_theme_file';
 
 import Footer from './Footer';
-import todos from '../todos';
-const { filters, TodoItem } = todos;
+import TodoItem from './TodoItem';
+
+import { filters } from '../constants';
+import * as allActions from '../actions';
 
 const defaultStyle = {
   width: 300,
@@ -34,6 +38,11 @@ class MainSection extends Component {
   getChildContext() {
     return { muiTheme: getMuiTheme(MyRawTheme) };
   }
+
+  componentDidMount() {
+    this.props.actions.fetch();
+  }
+
 
   handleClearCompleted() {
     const atLeastOneCompleted = this.props.todosData.some((todo) => todo.completed);
@@ -107,4 +116,19 @@ MainSection.propTypes = {
   actions: PropTypes.object.isRequired,
 };
 
-export default MainSection;
+function mapStateToProps(state) {
+  return {
+    todosData: state.todos,
+  };
+}
+
+function mapDispatchToProps(dispatch) {
+  return {
+    actions: bindActionCreators(allActions, dispatch),
+  };
+}
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(MainSection);
