+++ src/todos/index.js
@@ -1,20 +1,21 @@
-import { NAME, actionTypes, filters } from './constants';
+import { NAME, actionTypes } from './constants';
 import * as actions from './actions';
 import reducer from './reducer';
 import sagas from './sagas';
 
 // re-select
 // import * as selectors from './selectors'
-import TodoItem from './components/TodoItem.js';
-import TodoTextInput from './components/TodoTextInput.js';
+import MainSection from './components/MainSection';
+import TodoItem from './components/TodoItem';
+import TodoTextInput from './components/TodoTextInput';
 
 export default {
   NAME,
   actionTypes,
-  filters,
   actions,
   reducer,
   sagas,
+  MainSection,
   TodoItem,
   TodoTextInput,
 };
